+++ src/commands/mapshaper-proj.js
@@ -13,7 +13,7 @@ api.proj = function(dataset, destInfo, opts) {
       target = {},
       src, dest;
 
-  src = internal.getDatasetProjection(dataset);
+  src = internal.getDatasetCRS(dataset);
   if (!src) {
     stop("Unable to project -- source coordinate system is unknown");
   }
@@ -59,9 +59,9 @@ api.proj = function(dataset, destInfo, opts) {
 
 
 // @source: a layer identifier, .prj file or projection defn
-// Converts layer ids and .prj files to projection defn
+// Converts layer ids and .prj files to CRS defn
 // Returns projection defn
-internal.getProjectionInfo = function(name, catalog) {
+internal.getCrsInfo = function(name, catalog) {
   var dataset, sources, info = {};
   if (/\.prj$/i.test(name)) {
     dataset = api.importFile(name, {});
@@ -73,7 +73,7 @@ internal.getProjectionInfo = function(name, catalog) {
     sources = catalog.findCommandTargets(name);
     if (sources.length > 0) {
       dataset = sources[0].dataset;
-      info.crs = internal.getDatasetProjection(dataset);
+      info.crs = internal.getDatasetCRS(dataset);
       info.prj = dataset.info.prj; // may be undefined
       // defn = internal.crsToProj4(P);
     } else {